Extract TimerState and UpdateTask
More nested classes moved into their own file. Again, this should make things more readable and better testable.
