+++ app/src/main/java/com/apps/adrcotfas/goodtime/TimerState.java
@@ -1,0 +1,10 @@
+package com.apps.adrcotfas.goodtime;
+
+public enum TimerState {
+    INACTIVE,
+    ACTIVE_WORK,
+    PAUSED_WORK,
+    ACTIVE_BREAK,
+    FINISHED_WORK,
+    FINISHED_BREAK
+}
